Possible to run validate on multiple levels into the same df

+++ validate.py
@@ -28,26 +28,31 @@ def main(args):
 
         stats_gt = run_validate(os.path.join(result_path, gt_file),
                                 args.validation_correct_file,
-                                args.validate_path)
+                                args.validate_path, suffix=args.name)
         stats_regular = run_validate(os.path.join(result_path, 
                                                   regular_clustering_file),
                                      args.validation_correct_file,
-                                     args.validate_path)
+                                     args.validate_path, suffix=args.name)
         columns_different = ["Rand", "Rec.", "M", "N", "S", "Prec.", "AdjRand", "NMI"]
+        columns_different = [col + args.name for col in columns_different]
         for column in columns_different:
             stats_regular[column + "_gt"] = stats_gt[column]
-        stats_regular["gt"] = threshold
         
         stats[result_dir] = stats_regular
-
     stats_df = p.DataFrame.from_dict(stats, orient="index")
-    stats_df.to_csv(args.output_file)
+    if os.path.isfile(args.output_file) and args.merge_files:
+        original_df = p.io.parsers.read_csv(args.output_file, index_col=0)
+        original_df = original_df.join(stats_df, how='outer')
+        original_df.to_csv(args.output_file)
+    else:
+        stats_regular["gt"] = threshold
+        stats_df.to_csv(args.output_file)
 
-def run_validate(cfile, sfile, validate_path):
+def run_validate(cfile, sfile, validate_path, suffix=""):
     validate_so = subprocess.check_output(['perl', validate_path, 
                                            '--sfile={}'.format(sfile),
                                            '--cfile={}'.format(cfile) ])
-    headers = validate_so.split('\n')[0].split('\t')
+    headers = [h + suffix for h in validate_so.split('\n')[0].split('\t')]
     stats = validate_so.split('\n')[1].split('\t')
     stats_dict = dict(zip(headers, stats))
 
@@ -61,5 +66,8 @@ if __name__=="__main__":
                               "all contigs"))
     parser.add_argument('output_file', help="File where all contigs end up.")
     parser.add_argument('validate_path', help="Path to validation script.")
+    parser.add_argument('--name', default="", help="Name of run, to be appended to table columns.") 
+    parser.add_argument('--merge_files', action='store_true',
+            help="Merge with existing output file, otherwise overwrite")
     args = parser.parse_args()
     main(args)
